+++ benefits/RELEASES.md
@@ -150,3 +150,13 @@ Based on:
 - [python v4.2.0] benefits
 ### Releases
 - [PyPI v4.2.0] https://pypi.org/project/wingspan_benefits/4.2.0 - benefits
+
+## 2024-03-18 01:25:33
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.1] benefits
+### Releases
+- [PyPI v4.2.1] https://pypi.org/project/wingspan_benefits/4.2.1 - benefits